% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{gc_epub}
\alias{gc_epub}
\title{Creates an R Markdown epub Thesis document}
\usage{
gc_epub(...)
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}
}
\value{
A ebook version of the gc
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a epub version of the gc.
}
\examples{
\dontrun{
output:gcdown::gc_epub
}
}
