% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{gc_gitbook}
\alias{gc_gitbook}
\title{Creates an R Markdown gitbook Thesis document}
\usage{
gc_gitbook(...)
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}
}
\value{
A gitbook webpage
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a webpage version of the gc.
}
\examples{
\dontrun{
output:gcdown::gc_gitbook
}
}
