% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesisdown.R
\docType{package}
\name{gcdown}
\alias{gcdown}
\alias{gcdown-package}
\title{thesisdown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
Using the bookdown package and LaTeX and Word thesis templates from Reed College, this package aims to encourage reproducible research using simple Markdown syntax while embedding all of the R code to produce plots and analyses as well. This package works particularly for dissertations for the City University of New York's Graduate Center.
}
\section{gc_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{gc_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

\section{gc_epub}{

Creates an R Markdown thesis template as an ebook
}

\author{
\strong{Maintainer}: Chester Ismay \email{chester.ismay@gmail.com}

Authors:
\itemize{
  \item Nick Solomon \email{nick@nicksolomon.me}
  \item Sebastian Hoyos-Torres \email{sebtorres7021@gmail.com}
}

Other contributors:
\itemize{
  \item Shir Dekel \email{shirdekel@yahoo.com.au} [contributor]
}

}
